+++ actiontext/lib/action_text/engine.rb
@@ -60,6 +60,10 @@ module ActionText
           "[#{caption || filename}]"
         end
 
+        def attachable_markdown_representation(caption = nil)
+          "[#{caption || filename}]"
+        end
+
         def to_trix_content_attachment_partial_path
           to_editor_content_attachment_partial_path
         end